Virtual kubelet provider moved to pkg
